Refactors drawing function
Was working on the shader function and was struggling very hard to
get the functionality I wanted out of it. After a lot of messing around
I decided that my problem was how messy my draw function was. It was
doing too many things. I've separated the draw function into the click
part and the hover with mouse pressed down part. Those functions now
get called in a separate function dedicated to setting the event
listeners. After this change, refactored the rest of my functions to
work with the changes.