Creates drawing modifier buttons
All of the button code was going to be relatively similar so I deleted
the individual buttons and created a generic modifier button function
that takes a class name as a parameter. It also only allows one button
to be toggled at a time. Except for the grid line button which works
independently.